refactor: 테스트 헬퍼 및 coupon 테스트 V2 참조 제거
